+++ registry/default/blocks/data-grid-keybindings/data-grid-keybindings.ts
@@ -2,5 +2,6 @@
 
 export { DataGridKeybindingsDialog, type DataGridKeybindingsDialogProps } from "./keybindings-dialog";
 export { DataGridKeybindingsShortcut, type DataGridKeybindingsShortcutProps } from "./keybindings-shortcut";
+export { BindingChips, bindingTokens } from "./binding-label";
 export { KEYBINDING_CATEGORIES, ACTION_CATEGORY, categoryForAction, categoryLabel, type KeybindingCategory } from "./action-groups";
 export { labelForAction } from "./action-labels";